Create connection between exercises and workouts, add toggle button to hide/open full info

+++ app/src/main/java/com/example/samurairoad/room/tables/ExerciseTableModel.kt
@@ -30,5 +30,5 @@ data class ExerciseTableModel(
 ){
     @PrimaryKey(autoGenerate = true)
     @ColumnInfo(name = "id")
-    var Id: Int? = null
+    var Id: Long? = null
 }
