+++ app/src/main/java/com/example/samurairoad/room/tables/WorkoutTableModel.kt
@@ -15,5 +15,5 @@ data class WorkoutTableModel(
 
     @PrimaryKey(autoGenerate = true)
     @ColumnInfo(name = "id")
-    var Id: Int? = null
+    var Id: Long? = null
 }
